trocando_conveyor_por_storage
