+++ 2022-RapidReact/src/main/java/frc/robot/commands/commandgroups/AutoRoutine.java
@@ -7,7 +7,7 @@ package frc.robot.commands.commandgroups;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants;
-import frc.robot.commands.auto.AutoConveyor;
+import frc.robot.commands.auto.AutoStorage;
 import frc.robot.commands.auto.AutoIntake;
 import frc.robot.commands.auto.AutoShoot;
 import frc.robot.commands.auto.DriveStraight;
@@ -27,13 +27,13 @@ public class AutoRoutine extends ParallelCommandGroup {
       //new AutoNoShooter(storageSystem, shooter, driveTrain, intake),
 
       new SequentialCommandGroup(
-        new AutoConveyor(storageSystem, Constants.Storage.SPEED, 4),
+        new AutoStorage(storageSystem, Constants.Storage.SPEED, 4),
         new ParallelCommandGroup(
           new AutoIntake(intake, 4),
           new DriveStraight(driveTrain, 0.6, 3.5)
         ),
         new DriveStraight(driveTrain, -0.7, 3),
-        new AutoConveyor(storageSystem, Constants.Storage.SPEED, 4)
+        new AutoStorage(storageSystem, Constants.Storage.SPEED, 4)
       ),
       new AutoShoot(shooter, -Constants.Shooter.SPEED, time)
     );
